requirements.txt
django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.27
firebase-admin>=6.4
python-decouple>=3.8
django-cors-headers>=4.3
requests>=2.31
cloudinary>=1.36
google-genai>=1.0
